chore(refactor): replace utils, helpers, constant folder into shared folder
