+++ middleware/pagination/paginationValidation.js
@@ -1,5 +1,5 @@
 /* eslint-disable consistent-return */
-const { responseData } = require('../../helpers/responseDataHelper');
+const { responseData } = require('../../shared-v1/helpers/responseDataHelper');
 
 let errorResponse = {};
 const maxPageSizeValidation = (req, res, next) => {
