+++ modules/employee/employee.controller.js
@@ -1,6 +1,6 @@
-const { generateToken } = require('../../helpers/jwtHelper');
-const { passwordHashing } = require('../../helpers/passwordHelper');
-const { responseData } = require('../../helpers/responseDataHelper');
+const { generateToken } = require('../../shared-v1/helpers/jwtHelper');
+const { passwordHashing } = require('../../shared-v1/helpers/passwordHelper');
+const { responseData } = require('../../shared-v1/helpers/responseDataHelper');
 const {
   createEmployee,
   findEmployeeById,
