+++ shared-v1/helpers/jwtHelper.js
@@ -1,8 +1,8 @@
 const jwt = require('jsonwebtoken');
 const dotenv = require('dotenv');
 const { jwtDecode } = require('jwt-decode');
-const { findUserById } = require('../modules/user/user.service');
-const { findEmployeeById } = require('../modules/employee/employee.service');
+const { findUserById } = require('../../modules/user/user.service');
+const { findEmployeeById } = require('../../modules/employee/employee.service');
 
 dotenv.config();
