+++ shared-v1/helpers/responseDataHelper.js
